integrate age more directly
before, we got age when constructing persona from birthyr but an artifact of doing this is that users could not search by age since it was not a column in the dataframe. So in this commit, I made it so upon reading in ANES we compute age which then allows it to be searched like other attributes
Also updated the mapping, print_anes_values, and the persona creation to reflect this. Added unit tests for a variety of ages.